requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
motor==3.3.2
pymongo==4.5.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Earth Engine & Geospatial
earthengine-api==0.1.370
geemap==0.28.0
geopandas==0.13.2
shapely==2.0.1

# Data Processing
pandas==2.1.3
numpy==1.24.3
matplotlib==3.7.2

# PDF Generation
reportlab==4.0.4

# HTTP Clients
requests==2.31.0
aiofiles==23.2.1

# Environment & Config
python-dotenv==1.0.0
pydantic[email]==2.4.2
pydantic-settings==2.1.0

# Rate Limiting
slowapi==0.1.9

# Retry Logic
tenacity==8.2.3

# Async Utilities
asgiref==3.7.2

# Development
python-dateutil==2.8.2